Extract calendar state so it can be shared between components
