+++ pages/calendar.js
@@ -14,98 +14,19 @@ import {
 import { getAuth, onAuthStateChanged } from "firebase/auth";
 import { useEffect } from "react";
 
+import useWeekStore from "lib/state";
+
 const localizedFormat = require("dayjs/plugin/localizedFormat");
 dayjs.extend(localizedFormat);
 
-// Day is a dayjs() date object, direction is either "next" or "previous"
-function generateWeek(day, direction = "") {
-  // Find nearest monday and start week array with that monday
-  let j = day;
-  // Max number of days from Monday is 6 (Sunday)
-  for (let i = 0; i < 6; i++) {
-    if (j.format("dddd") === "Monday") {
-      break;
-    }
-    j = j.subtract(1, "day");
-  }
-
-  // Adjust week offset if necessary
-  if (direction === "next") {
-    j = j.add(7, "day");
-  } else if (direction === "previous") {
-    j = j.subtract(7, "day");
-  }
-
-  // Generate week array with all 7 days starting on a given day
-  let week = [];
-  for (let i = 0; i < 7; i++) {
-    week[i] = j;
-    j = j.add(1, "day");
-  }
-
-  return week;
-}
-
-// This feels hacky but for some reason there's no way to set a default value for weekdays based on startDay since startDay doesn't exist when I'm trying to use it
-const startDayInitial = dayjs();
-const weekdaysInitial = generateWeek(startDayInitial, "current");
-const useWeekStore = create((set) => ({
-  weekdays: weekdaysInitial,
-  startDay: startDayInitial,
-  calendar: [],
-  setWeek: (day, direction) => {
-    set((state) => ({
-      weekdays: generateWeek(day, direction),
-    }));
-  },
-  setStart: (day) => {
-    set((state) => ({
-      startDay: day,
-    }));
-  },
-  setCalendar: (calendar) => {
-    set((state) => ({
-      calendar: calendar,
-    }));
-  },
-}));
-
 const Calendar = (props) => {
-  const startDay = useWeekStore((state) => state.startDay);
   const weekdays = useWeekStore((state) => state.weekdays);
   const calendar = useWeekStore((state) => state.calendar);
   const setWeek = useWeekStore((state) => state.setWeek);
-  const setStart = useWeekStore((state) => state.setStart);
-  const setCalendar = useWeekStore((state) => state.setCalendar);
+  const updateCalendar = useWeekStore((state) => state.updateCalendar);
 
-  console.log("testABC");
   useEffect(() => {
-    setWeek(startDay, "current");
-
     // Gather initial calendar data on component load
-    const updateCalendar = async () => {
-      const auth = getAuth();
-      onAuthStateChanged(auth, async (user) => {
-        const startDate = dayjs().format();
-        const token = getAuth().currentUser.accessToken;
-        console.log(token);
-        const response = await fetch("/api/getCalendarData", {
-          method: "POST",
-          headers: {
-            authorization: token,
-          },
-          body: JSON.stringify({
-            startDate: startDate,
-            numberOfDays: 7,
-          }),
-        });
-        const data = await response.json();
-        const newCalendar = data.calendar;
-        console.log(newCalendar);
-        setCalendar(newCalendar);
-      });
-    };
-
     updateCalendar();
   }, []);
 
@@ -140,8 +61,8 @@ const Calendar = (props) => {
                       size={48}
                       className="border rounded-lg cursor-pointer mr-1"
                       onClick={() => {
-                        setWeek(startDay, "previous");
-                        setStart(startDay.subtract(7, "day"));
+                        setWeek(weekdays[0], "previous");
+                        updateCalendar();
                       }}
                     />
                     <IconChevronRight
@@ -149,8 +70,8 @@ const Calendar = (props) => {
                       size={48}
                       className="border rounded-lg cursor-pointer mr-2"
                       onClick={() => {
-                        setWeek(startDay, "next");
-                        setStart(startDay.add(7, "day"));
+                        setWeek(weekdays[0], "next");
+                        updateCalendar();
                       }}
                     />
                   </Box>
@@ -159,9 +80,7 @@ const Calendar = (props) => {
                     size={48}
                     className="border rounded-lg cursor-pointer mr-2"
                     onClick={() => {
-                      const updated = dayjs();
-                      setStart(updated);
-                      setWeek(updated);
+                      updateCalendar();
                     }}
                   />
                 </Box>
